new files

+++ src/components/Passwords.tsx
@@ -1,27 +1,39 @@
-import { useEffect, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import Copy from '../assets/Copy';
+import EyeClose from '../assets/EyeClose';
+import EyeOpen from '../assets/EyeOpen';
 import toast, { Toaster } from 'react-hot-toast';
 import Edit from '../assets/Edit';
 import Delete from '../assets/Delete';
 import { Tooltip } from 'react-tooltip';
 import { useNavigate } from 'react-router-dom';
+import { FormContext } from '../context/FormContext';
 
 const Passwords = () => {
   const navigate = useNavigate();
+  const formContext = useContext(FormContext);
   const location = useLocation();
   const [passwords, setPasswords] = useState<any[]>([]);
+  const [pass_show, setPass_show] = useState(false);
+
+  if (!formContext) return null;
+  const { setForm } = formContext;
 
   useEffect(() => {
     const savedPasswords = JSON.parse(localStorage.getItem("password") || "[]");
     setPasswords(savedPasswords);
   }, [location.key]);
 
+  const togglePassword = () => {
+    setPass_show(!pass_show)
+  }
+
   const notify = (popup: any,) => toast(<h6>{popup}</h6>, {
     duration: 2000,
     position: 'top-center',
     style: { "height": "40px", "width": "max-contnet" },
-    className: '!bg-transparent !decoration-0 !shadow !font-medium',
+    className: '!bg-white !decoration-0 !shadow !font-medium',
     icon: "✅",
     iconTheme: {
       primary: '#000',
@@ -34,72 +46,92 @@ const Passwords = () => {
     removeDelay: 1000,
   });
   console.log(passwords);
-  
+
   const handleCopytext = (text: any, field: any) => {
     navigator.clipboard.writeText(text);
     notify(`Copied ${field} to ClipBoard!`)
   }
 
-  const handleEdit = (id: any) => {
-    console.log('Id Edited:- ',id);
+
+  const handleEdit = (item: any) => {
+    let ok = confirm("Do you want to Edit this Field? (You have to edit thit Field once you marked for edit!)")
+    if (ok) {
+
+      const data = {
+        url: item.url,
+        username: item.username,
+        password: item.password
+      }
+      setForm(data);
+      console.log('data Edited:- ', item.id);
+      console.log('Id deleted:- ', item.id);
+      passwords.forEach((e, i) => {
+        if (e.id === item.id) {
+          passwords.splice(i, 1);
+        }
+      });
+      setPasswords(passwords)
+      localStorage.setItem("password", JSON.stringify(passwords))
+      navigate("/")
+    }
 
   }
+
+
   const handleDelete = (id: any) => {
-    console.log('Id deleted:- ',id);
-    console.log(passwords);
-    passwords.forEach((e,i) => {
-      if (e.id===id) {
-        passwords.splice(i, 1);
-      }
-    });
-    console.log(passwords);
-    
-    // const filterList:any[] = passwords.filter((item)=>{
-    //   item.id!==id
-    // })
-    
-    setPasswords(passwords)
-    localStorage.setItem("password",JSON.stringify(passwords))
-    navigate("/Passwords")
-  
-    
+    let ok = confirm("Do you want to Delete this Field? (This Action is Irreversible! )")
+    if (ok) {
+      console.log('Id deleted:- ', id);
+      passwords.forEach((e, i) => {
+        if (e.id === id) {
+          passwords.splice(i, 1);
+        }
+      });
+      setPasswords(passwords)
+      localStorage.setItem("password", JSON.stringify(passwords))
+      setTimeout(() => {
+        navigate("/Passwords");
+      }, 1000);
+      notify("Field Deleted!")
+    }
   }
   const handleDeleteAll = () => {
     if (passwords.length > 0) {
-      localStorage.clear()
-      navigate("/passwords")
+      let ok = confirm("Do you want to Delete All Fields? (This Action is Irreversible! )")
+      if (ok) {
+        localStorage.clear()
+        setTimeout(() => {
+          navigate("/Passwords");
+        }, 1000);
+        notify("All Fields Deleted!")
+      }
     }
 
-
   }
-
-
-
-
-
   return (
     <>
-      <div className="passwords select-none mt-10 container m-auto h-screen">
+      <div className="passwords p-5 text-xs sm:text-sm md:text-lg select-none mt-10 container m-auto h-screen overflow-auto">
         <div className="relative overflow-x-auto rounded-md shadow-md">
-          <table className="w-full text-sm text-left rtl:text-right text-gray-400 table-fixed">
+          <table className="w-full text-left rtl:text-right text-gray-400 table-fixed">
             <colgroup>
               <col className="w-1/4" />
               <col className="w-1/4" />
               <col className="w-1/4" />
               <col className="w-1/4" />
             </colgroup>
-            <thead className="text-lg text-black uppercase bg-slate-300">
+
+            <thead className="h-10  text-black uppercase bg-slate-300">
               <tr className="text-center">
-                <th scope="col" className="px-6 py-3">Website</th>
-                <th scope="col" className="px-6 py-3">Username</th>
-                <th scope="col" className="px-6 py-3">Password</th>
-                <th scope="col" className="px-6 py-3">
+                <th scope="col" className="md:px-6 md:py-3">Website</th>
+                <th scope="col" className="md:px-6 md:py-3">Username</th>
+                <th scope="col" className="md:px-6 md:py-3">Password</th>
+                <th scope="col" className="md:px-6 md:py-3 max-w-[120px]">
                   <div className="flex items-center justify-center gap-10">
                     <span>Actions</span>
                     <span onClick={() => handleDeleteAll()} data-tooltip-id="Dltall"
-                      className="show-password flex items-center justify-center h-10 w-28 rounded-md shadow-md">
+                      className="show-password flex items-center justify-center md:h-10 md:w-28 h-5 w-14 rounded-md shadow-md">
                       <button
-                        className="flex justify-center items-center w-max h-max text-black transition-colors duration-200 bg-transparent border border-black rounded-md focus:outline-none cursor-pointer hover:bg-slate-300 hover:scale-105 text-[13px]"
+                        className="flex justify-center items-center w-max h-max text-black transition-colors duration-200 bg-transparent border border-black rounded-md focus:outline-none cursor-pointer hover:bg-slate-300 hover:scale-105 text-[10px] md:text-[13px]"
                       >
                         <span className="font-semibold">Delete All Passwords!</span>
                       </button>
@@ -114,7 +146,7 @@ const Passwords = () => {
                 passwords.map((item, index: number) => (
                   <tr
                     key={index}
-                    className="text-center bg-slate-400 border-b text-lg border-slate-600 hover:bg-slate-500"
+                    className="text-center bg-slate-400 border-b  text-xs sm:text-sm md:text-lg border-slate-600 hover:bg-slate-500"
                   >
 
                     <td className="px-6 py-2 text-black align-top">
@@ -128,6 +160,7 @@ const Passwords = () => {
                           {item.url}
                         </a>
                         <span onClick={() => handleCopytext(item.url, "URL")} className="mt-1"><Copy /></span>
+
                         <Toaster />
                       </div>
                     </td>
@@ -144,8 +177,18 @@ const Passwords = () => {
 
                     <td className="px-6 py-2 text-black align-top">
                       <div className="flex items-center justify-center gap-2 break-all whitespace-normal">
-                        <span className="break-all">{item.password}</span>
-                        <span onClick={() => handleCopytext(item.password, "Password")} className="mt-1"><Copy /></span>
+                        <span style={{ WebkitTextSecurity: pass_show ? 'none' : 'disc' } as React.CSSProperties} className="break-all">  {pass_show ? item.password : '●'.repeat(8)}</span>
+                        <span className="relative top-[3px] flex items-center justify-center">
+                          {pass_show ?
+                            <button onClick={togglePassword}><EyeClose /></button>
+
+                            :
+                            <button onClick={togglePassword}> <EyeOpen /></button>
+
+                          }
+                        </span>
+                        {pass_show && <span onClick={() => handleCopytext(item.password, "Password")} className="mt-1"><Copy /></span>}
+
                         <Toaster />
                       </div>
                     </td>
@@ -153,7 +196,9 @@ const Passwords = () => {
 
                     <td className="px-6 py-2 text-black align-top">
                       <div className="flex items-center justify-center gap-2">
-                        <span onClick={() => handleEdit(item.id)} className="cursor-pointer"><Edit /></span>
+                        <span onClick={() => handleEdit(item)}
+
+                          className="cursor-pointer"><Edit /></span>
                         <span onClick={() => handleDelete(item.id)} className="cursor-pointer"><Delete /></span>
                         <Toaster />
                       </div>
@@ -170,7 +215,7 @@ const Passwords = () => {
             </tbody>
           </table>
         </div>
-      </div>
+      </div> <br /> <br />
     </>
   )
 }
